add notes from post-install steps into node_modules
this is currently needed becuase directly importing the code
from another typescript project leads to this error

```
Cannot use import statement outside a module
```

whereas it's possible to use compiled code in `dist` -- but
you need to compile it. There's likely an automated way to
do this but I just freaking want to import the lib now.